Summarising the results of the model fitting functions

S3 method for class 'sars'. summary.sars creates summary statistics for objects of class 'sars'. The exact summary statistics computed depends on the 'Type' attribute (e.g. 'multi') of the 'sars' object. The summary method generates more useful information for the user than the standard model fitting functions. Another S3 method (print.summary.sars; not documented) is used to print the output.

# S3 method for sars
summary(object, ...)

Arguments

object
An object of class 'sars'.

…
Further arguments.

Value

The summary.sars function returns an object of class "summary.sars". A print function is used to obtain and print a summary of the model fit results.

For a 'sars' object of Type 'fit', a list with 16 elements is returned that contains useful information from the model fit, including the model parameter table (with t-values, p-values and confidence intervals), model fit statistics (e.g. R2, AIC), the observed shape of the model and whether or not the fit is asymptotic, and the results of any additional model checks undertaken (e.g. normality of the residuals).

For a 'sars' object of Type 'multi', a list with 4 elements is returned: (i) a vector of the names of the models that were successfully fitted and passed any additional checks, (ii) a character string containing the name of the criterion used to rank models, (iii) a data frame of the ranked models, and (iv) a vector of the names of any models that were not fitted or did not pass any additional checks. In regards to (iii; Model_table), the dataframe contains the fit summaries for each successfully fitted model (including the value of the model criterion used to compare models, the R2 and adjusted R2, and the observed shape of the fit); the models are ranked in decreasing order of information criterion weight.

For a 'sars' object of Type 'lin_pow', a list with up to 7 elements is returned: (i) the model fit output from the lm function, (ii) the fitted values of the model, (iii) the observed data, (iv and v) the results of the residuals normality and heterogeneity tests, and (vi) the log-transformation function used. If the argument compare = TRUE is used in lin_pow, a 7th element is returned that contains the parameter values from the non-linear power model.

Examples

data(galap)
#fit a multimodel SAR and get the model table
mf <- sar_average(data = galap)

#> 
#>  Now attempting to fit the 20 SAR models: 
#> 
#> --  multi_sars --------------------------- multi-model SAR --
#> > power    : v
#> > powerR   : v
#> > epm1     : v
#> > epm2     : v
#> > p1       : v
#> > p2       : v
#> > loga     : v
#> > koba     : v
#> > mmf      : v
#> > monod    : v
#> > negexpo  : v
#> > chapman  : Warning: could not compute parameters statistics
#> > weibull3 : v
#> > asymp    : v
#> > ratio    : v
#> > gompertz : v
#> > weibull4 : v
#> > betap    : v
#> > heleg    : v
#> > linear   : v
#> 
#> Model fitting completed - all models succesfully fitted.  Now undertaking model validation checks.
#> Additional models  will be excluded if necessary:

#> 
#> 4 models failed the residuals normality test and  have been excluded  from the multi SAR:
#> Extended Power model 1, Asymptotic regression, Cumulative Weibull 4 par., Linear model

#> 16 remaining models used to construct the multi  SAR:
#>  Power, PowerR, Extended Power model 2, Persistence function 1, Persistence function 2, Logarithmic, Kobayashi, MMF, Monod, Negative exponential, Chapman Richards, Cumulative Weibull 3 par., Rational function, Gompertz, Beta-P cumulative, Heleg(Logistic) 
#> -------------------------------------------------------------

summary(mf)

#> 
#> Sar_average object summary:
#> 
#> 16  models successfully fitted
#> 
#> The following models could not be fitted or were  removed due to model checks:
#> Extended Power model 1, Asymptotic regression, Cumulative Weibull 4 par., Linear model 
#> 
#> Ranked models based on AICc  weights:
#> 
#>       Model Weight    AICc    R2   R2a     Shape Asymptote
#> 1   negexpo  0.219 186.832 0.557 0.488 convex up      TRUE
#> 2     monod  0.178 187.246 0.545 0.475 convex up      TRUE
#> 3      koba  0.118 188.073 0.521 0.447 convex up     FALSE
#> 4     power  0.073 189.031 0.491 0.413 convex up     FALSE
#> 5      loga  0.067 189.191 0.486 0.407 convex up     FALSE
#> 6        p1  0.066 189.246 0.589 0.486 convex up     FALSE
#> 7  gompertz  0.061 189.393 0.585 0.482 convex up      TRUE
#> 8  weibull3  0.038 190.324 0.561 0.451 convex up      TRUE
#> 9   chapman  0.036 190.468 0.557 0.446 convex up      TRUE
#> 10    ratio  0.035 190.494 0.556 0.445 convex up      TRUE
#> 11    heleg  0.029 190.882 0.545 0.431 convex up      TRUE
#> 12      mmf  0.029 190.882 0.545 0.431 convex up      TRUE
#> 13       p2  0.019 191.712 0.521 0.401 convex up     FALSE
#> 14   powerR  0.014 192.283 0.503 0.379 convex up     FALSE
#> 15     epm2  0.014 192.329 0.502 0.377   sigmoid     FALSE
#> 16    betap  0.002 196.018 0.522 0.349 convex up     FALSE

summary(mf)$Model_table

#>       Model Weight    AICc    R2   R2a     Shape Asymptote
#> 1   negexpo  0.219 186.832 0.557 0.488 convex up      TRUE
#> 2     monod  0.178 187.246 0.545 0.475 convex up      TRUE
#> 3      koba  0.118 188.073 0.521 0.447 convex up     FALSE
#> 4     power  0.073 189.031 0.491 0.413 convex up     FALSE
#> 5      loga  0.067 189.191 0.486 0.407 convex up     FALSE
#> 6        p1  0.066 189.246 0.589 0.486 convex up     FALSE
#> 7  gompertz  0.061 189.393 0.585 0.482 convex up      TRUE
#> 8  weibull3  0.038 190.324 0.561 0.451 convex up      TRUE
#> 9   chapman  0.036 190.468 0.557 0.446 convex up      TRUE
#> 10    ratio  0.035 190.494 0.556 0.445 convex up      TRUE
#> 11    heleg  0.029 190.882 0.545 0.431 convex up      TRUE
#> 12      mmf  0.029 190.882 0.545 0.431 convex up      TRUE
#> 13       p2  0.019 191.712 0.521 0.401 convex up     FALSE
#> 14   powerR  0.014 192.283 0.503 0.379 convex up     FALSE
#> 15     epm2  0.014 192.329 0.502 0.377   sigmoid     FALSE
#> 16    betap  0.002 196.018 0.522 0.349 convex up     FALSE

#Get a summary of the fit of the linear power model
fit <- lin_pow(galap, con = 1, compare = TRUE)
summary(fit)

#> Model = Log-log power
#> 
#> Log-transformation function used: log()
#> 
#> Call:
#> lm(formula = logT(S) ~ logT(A), data = data)
#> 
#> Residuals:
#>     Min      1Q  Median      3Q     Max 
#> -1.3591 -0.7584  0.1177  0.6009  1.0739 
#> 
#> Coefficients:
#>      Estimate Std. Error t value Pr(>|t|)    
#> LogC  3.01865    0.35442   8.517 6.56e-07 ***
#> z     0.33854    0.08523   3.972  0.00139 ** 
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> 
#> Residual standard error: 0.7626 on 14 degrees of freedom
#> Multiple R-squared:  0.5298,	Adjusted R-squared:  0.4962 
#> F-statistic: 15.78 on 1 and 14 DF,  p-value: 0.001391
#> 
#> Power (non-linear) parameters:
#>  c = 33.18 
#>  z = 0.28